Backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.1.2
python-dotenv==1.0.0
openai==0.28.1
nltk==3.8.1
PyJWT==2.7.0
